+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Changes:
 
+* feature: add option to run dataplane as a gateway without inbounds
+  [#503](https://github.com/Kong/kuma/pull/503)
 * feature: add `METRICS` column to the table output of `kumactl get meshes` to make it visible whether Prometheus settings have been configured
   [#502](https://github.com/Kong/kuma/pull/502)
 * feature: automatically set default values for Prometheus settings in the Mesh resource